fixato layout vista a lista, creato componente per il btn della wishlist per evitare ripetizioni di codice, fixata background img della home
