+++ src/components/PLCardList.jsx
@@ -1,90 +1,51 @@
 import { Link } from "react-router";
-import { useOutletContext } from "react-router-dom";
+//import { useOutletContext } from "react-router-dom";
 //icons
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faDice } from "@fortawesome/free-solid-svg-icons";
-
+import BtnWishlist from "./BtnWishlist";
 export default function PLCardLIst({ game }) {
-  const { wishlist, setWishlist, triggerAlert } = useOutletContext();
-  const toggleWishlist = () => {
-    const isInWishlist = wishlist.some((p) => p.id === game.id);
-
-    if (isInWishlist) {
-      setWishlist(wishlist.filter((p) => p.id !== game.id));
-    } else {
-      setWishlist([...wishlist, game]);
-    }
-  };
-
-  // Verifico se il prodotto è nei preferiti per cambiare l’icona
-  const isFavorite = wishlist.some((p) => p.id === game.id);
   return (
     <div className="card mb-4">
-      <div
-        key={game.id}
-        className="cardlist  card-body text-dark d-flex align-items-center justify-content-between flex-wrap"
-      >
-        <div className=" ">
-          <h6>{game.name}</h6>
-          <p>{game.description}</p>
-        </div>
-        <div className="d-flex list-group-flush">
-          <div className="list-group-item">
-            <span className="badge bg-secondary me-2">
-              Durata: {game.duration}'
-            </span>
+      <div className="cardlist  card-body text-dark">
+        <div className="py-2">
+          <div className="me-5 ">
+            <h6>{game.name}</h6>
+            <p className="">{game.description}</p>
           </div>
-          <div className="list-group-item ml-2">
-            <span className="badge bg-secondary me-2">Età: {game.age}+</span>
-          </div>
-          <div className="list-group-item ml-2">
-            <span className="badge bg-secondary me-2">
-              Giocatori: {game.players}
-            </span>
-          </div>
-          <div className="list-group-item ml-2">
-            <span className="badge bg-success text-dark me-2">
-              {game.difficulty}
-            </span>
+          <div className="d-flex list-group-flush  ">
+            <div className="list-group-item">
+              <span className="badge bg-black me-2">
+                Durata: {game.duration}'
+              </span>
+            </div>
+            <div className="list-group-item ml-2">
+              <span className="badge bg-black me-2">Età: {game.age}+</span>
+            </div>
+            <div className="list-group-item ml-2">
+              <span className="badge bg-black me-2">
+                Giocatori: {game.players}
+              </span>
+            </div>
+            <div className="list-group-item ml-2">
+              <span className="badge bg-success text-white me-2">
+                {game.difficulty}
+              </span>
+            </div>
           </div>
         </div>
 
-        <div>
-          <Link to={`/products/${game.slug}`} className="btn btn-primary">
-            Vai al gioco
-            <FontAwesomeIcon icon={faDice} />
-          </Link>
-          <button
-            data-bs-toggle="tooltip"
-            data-bs-placement="top"
-            type="button"
-            className="btn text-danger border-0"
-            onClick={() => {
-              toggleWishlist();
-              triggerAlert(
-                isFavorite
-                  ? `${game.name} Rimosso dai preferiti`
-                  : ` ${game.name} Aggiunto ai preferiti`,
-                isFavorite ? "danger" : "success"
-              );
-            }}
-            title={
-              isFavorite ? "Rimuovi dai preferiti" : "Aggiungi ai preferiti"
-            }
-          >
-            <i
-              className={
-                isFavorite
-                  ? "fi fi-ts-dice-d20 text-danger"
-                  : "fi fi-ts-dice-d20 text-black"
-              }
-              style={{
-                display: "inline-block",
-                fontSize: "30px",
-                animation: isFavorite ? "spin 0.5s ease-in-out" : "",
-              }}
-            ></i>
-          </button>
+        <div className="d-flex justify-content-between align-items-center  ">
+          <div>
+            <Link to={`/products/${game.slug}`} className="btn btn-primary">
+              Vai al gioco
+              <FontAwesomeIcon icon={faDice} />
+            </Link>
+          </div>
+          <div className="p-2">
+            {" "}
+            <BtnWishlist game={game} />
+          </div>
         </div>
       </div>
     </div>
